requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
google-api-core==2.11.1
google-auth==2.20.0
google-cloud-vision==3.4.3
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3
alembic==1.11.1
atomicwrites==1.4.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
build==0.10.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
coverage==7.2.7
cryptography==41.0.1
dnspython==2.3.0
dominate==2.8.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.4
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
more-itertools==9.1.0
packaging==23.1
passlib==1.7.4
pep517==0.13.0
Pillow==9.5.0
pip-review==1.3.0
pip-tools==6.13.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.1.0
pytest==7.3.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==2.0.16
tomli==2.0.1
visitor==0.1.3
wcwidth==0.2.6
Werkzeug==2.3.6
WTForms==3.0.1
